refactor(hydro_lang)!: move `backtrace` module under `ir`
It won't be used by most Hydro developers, and is only used in the IR, so better to group them.
