+++ hydro_lang/src/lib.rs
@@ -93,8 +93,6 @@ pub mod rewrites;
 
 mod staging_util;
 
-pub mod backtrace;
-
 #[cfg(feature = "deploy")]
 #[cfg_attr(docsrs, doc(cfg(feature = "deploy")))]
 pub mod test_util;
