+++ hydro_lang/src/location/mod.rs
@@ -12,9 +12,9 @@ use syn::parse_quote;
 use tokio_util::codec::{Decoder, Encoder, LengthDelimitedCodec};
 
 use super::builder::FlowState;
-use crate::backtrace::Backtrace;
 use crate::boundedness::Unbounded;
 use crate::cycle::{CycleCollection, ForwardRef, ForwardRefMarker};
+use crate::ir::backtrace::Backtrace;
 use crate::ir::{
     DebugInstantiate, HydroIrMetadata, HydroIrOpMetadata, HydroNode, HydroRoot, HydroSource,
 };
